Extract the framework initialization logic from the ContaoFramework class.
